commit 647709cb33adbc9ec046fb9ad41818ffc50b0bf3
    11735 cardbus: cast between incompatible function types
commit 7a2e057ded7c57de0de12bba546e48c437d4aba0
    11734 pcic: cast between incompatible function types
commit 16d00323713cfa922181df7d69d973398dea15ff
    11733 acpica: cast between incompatible function types
commit 660213e2089ab8409d54154d0a2987504a24fb23
    11732 portfs: cast between incompatible function types
commit 8ab009369fb9f9b1e690d195ddcfbd8165a53f34
    11731 kaio: cast between incompatible function types
commit 9dd95169794650baa0f951e24046175899beff70
    11730 shmsys: cast between incompatible function types
commit d0562c105d77a4d4da85007111e260da7ea2616f
    11754 ZoL trim port used wrong offset for sd unmap
commit aef5ddef46ea0369baf07aba4d38ed9b2cd8cd96
    11727 usb_ah: cast between incompatible function types
commit fe9b11956841aaade2b3d01981d7e372b9b613a5
    11728 exacctsys: cast between incompatible function types
commit 457cb919baede78b33c994ab8c7797c848513997
    11726 ttcompat: cast between incompatible function types
commit c6026c814721e22d6b8791b5bdfeeecb85ece59c
    11725 tirdwr: cast between incompatible function types
commit c4d76aa4b703811c0527424e3c63151eb77b9972
    11724 spppcomp: cast between incompatible function types
commit e81f5104293f278cbf9b239692af80ee10885830
    11723 cryptmod: cast between incompatible function types
commit bbe1f46d73a73e54a242bb59596e2daf4e619fe3
    11722 usbwcm: cast between incompatible function types
commit 700daa96bb3e99515f0abdc42f1904e78bc282da
    11658 Some capabilities are invalid for SMB 2.002
commit 4d633836fc25186ed4c118b2072d2b88cf87a700
    11720 rpcmod: cast between incompatible function types
commit 6caf82fa836a9721560a96669ccdd90b3fa7b644
    11719 timod: cast between incompatible function types
commit e10751a53e4fcc2850e57448ad13dc3caf40595e
    11755 loader: command_lsmod does show garbage on screen
commit eb62f060787d63335f2b0777ff86c16204bca93a
    10517 Convert dkio(7I) to mandoc
    10518 Convert fbio(7I) to mandoc
    10519 Convert fdio(7I) to mandoc
    10520 Convert hdio(7I) to mandoc
    10521 Convert iec61883(7I) to mandoc
    10522 Convert ipnat(7I) to mandoc
    10527 Convert mhd(7I) to mandoc
    10528 Convert mixer(7I) to mandoc
    10529 Convert mtio(7I) to mandoc
    10530 Convert prnio(7I) to mandoc
    10531 Convert quotactl(7I) to mandoc
    10532 Convert sesio(7I) to mandoc
    10533 Convert sockio(7I) to mandoc
    10560 Convert termio(7I) to mandoc
    10561 Convert termiox(7I) to mandoc
    10562 Convert visual_io(7I) to mandoc
    10563 Convert vt(7I) to mandoc
    10564 Convert uscsi(7I) to mandoc
    10596 Convert streamio(7I) to mandoc

+++ usr/src/uts/common/fs/portfs/port.c
@@ -388,7 +388,7 @@ static int64_t portfs(int, uintptr_t, uintptr_t, uintptr_t, uintptr_t,
 static struct sysent port_sysent = {
 	6,
 	SE_ARGC | SE_64RVAL | SE_NOUNLOAD,
-	(int (*)())portfs,
+	(int (*)())(uintptr_t)portfs,
 };
 
 static struct modlsys modlsys = {
@@ -404,7 +404,7 @@ portfs32(uint32_t arg1, int32_t arg2, uint32_t arg3, uint32_t arg4,
 static struct sysent port_sysent32 = {
 	6,
 	SE_ARGC | SE_64RVAL | SE_NOUNLOAD,
-	(int (*)())portfs32,
+	(int (*)())(uintptr_t)portfs32,
 };
 
 static struct modlsys modlsys32 = {
